Create Readme.md
# LeetCode Problem 572: Subtree of Another Tree

**Difficulty:** Easy

**Link to Problem:** [572. Subtree of Another Tree](https://leetcode.com/problems/subtree-of-another-tree/)

---

## Problem Description

Given the roots of two binary trees `root` and `subRoot`, return `True` if there is a subtree of `root` with the same structure and node values as `subRoot`, and `False` otherwise.

A subtree of a binary tree `tree` is a tree that consists of a node in `tree` and all of this node's descendants. The tree `tree` could also be considered as a subtree of itself.